# docs/swagger.yaml
definitions:
  main.EquipmentType:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  main.Genre:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample
  title: Your API Title
  version: "1.0"
paths:
  /equipment-types:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.EquipmentType'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить все типы оборудования
      tags:
      - equipment-types
    post:
      consumes:
      - application/json
      parameters:
      - description: Тип оборудования
        in: body
        name: equipment_type
        required: true
        schema:
          $ref: '#/definitions/main.EquipmentType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.EquipmentType'
        "400":
          description: Неверный запрос
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создать тип оборудования
      tags:
      - equipment-types
  /equipment-types/{id}:
    delete:
      parameters:
      - description: ID типа оборудования
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Удалено
          schema:
            type: string
        "404":
          description: Тип не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить тип оборудования
      tags:
      - equipment-types
    get:
      parameters:
      - description: ID типа оборудования
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.EquipmentType'
        "404":
          description: Тип оборудования не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить тип оборудования по ID
      tags:
      - equipment-types
    put:
      consumes:
      - application/json
      parameters:
      - description: ID типа оборудования
        in: path
        name: id
        required: true
        type: string
      - description: Обновлённые данные типа
        in: body
        name: equipment_type
        required: true
        schema:
          $ref: '#/definitions/main.EquipmentType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.EquipmentType'
        "400":
          description: Неверный JSON
          schema:
            type: string
        "404":
          description: Тип не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить тип оборудования
      tags:
      - equipment-types
  /genres:
    get:
      description: Возвращает список всех жанров
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Genre'
            type: array
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      summary: Получить все жанры
      tags:
      - genres
    post:
      consumes:
      - application/json
      description: Добавляет новый жанр в базу данных
      parameters:
      - description: Новый жанр
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/main.Genre'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Genre'
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Создать жанр
      tags:
      - genres
  /genres/{id}:
    delete:
      description: Удаляет жанр по ID
      parameters:
      - description: ID жанра
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Удалено
          schema:
            type: string
        "404":
          description: Жанр не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удалить жанр
      tags:
      - genres
    get:
      description: Возвращает жанр по его UUID
      parameters:
      - description: ID жанра
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Genre'
        "404":
          description: Жанр не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      summary: Получить жанр по ID
      tags:
      - genres
    put:
      consumes:
      - application/json
      description: Обновляет существующий жанр по ID
      parameters:
      - description: ID жанра
        in: path
        name: id
        required: true
        type: string
      - description: Обновленные данные жанра
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/main.Genre'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Genre'
        "400":
          description: Неверный JSON
          schema:
            type: string
        "404":
          description: Жанр не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Обновить жанр
      tags:
      - genres
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"